Fix MAE eval AMP dtype, loss_beta passthrough, and deprecated npho_branch fallback
- Add missing dtype=torch.bfloat16 to MAE eval autocast (was inconsistent with train path)
- Add beta parameter to get_pointwise_loss_fn and wire per-task loss_beta from regressor config
- Fix regressor npho_branch fallback from deprecated "relative_npho" to "npho"

+++ lib/engines/mae.py
@@ -737,7 +737,7 @@ def run_eval_mae(model, device, root_files, tree_name,
         profiler.stop()
 
         with torch.no_grad():
-            with torch.amp.autocast('cuda', enabled=amp):
+            with torch.amp.autocast('cuda', dtype=torch.bfloat16, enabled=amp):
                 # Get masked input for visualization (pass npho_threshold_norm for stratified masking)
                 profiler.start("forward")
                 x_masked, mask = model.random_masking(x_in, npho_threshold_norm=npho_threshold_norm)
